duplicate currency and date conversion modules and tests tidied up
