+++ app/processing/conversion/convert-currency.js
@@ -1,5 +1,5 @@
-function convertStringToPence (valueInPounds) {
-  const currencyArray = valueInPounds.split('.')
+const convertToPence = (valueInPounds) => {
+  const currencyArray = valueInPounds.toString().split('.')
   const pounds = currencyArray[0]
   const pence = (currencyArray[1] || '00').padEnd(2, '0')
   return Number(pounds + pence)
@@ -10,6 +10,6 @@ function convertToPounds (valueInPence) {
 }
 
 module.exports = {
-  convertStringToPence,
+  convertToPence,
   convertToPounds
 }
